pyproject.toml
[tool.poetry]
name = "openaiapidev"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.8"
python-dotenv = "^1.0.0"
langchain = "^0.0.206"
pandas = {extras = ["tabulate"], version = "^2.0.2"}
openpyxl = "^3.1.2"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
